backend/requirements.txt
blinker==1.8.2
click==8.1.7
Flask-SQLAlchemy==3.1.1
Flask==3.0.3
greenlet==3.0.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.2.4
python-dotenv==1.0.1
SQLAlchemy==2.0.30
typing_extensions==4.11.0
Werkzeug==3.0.3
